+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # v1.7.0-rc.17
 ## mm/dd/2020
 
+1. [](#improved)
+    * Fall back through various templates scenarios if they don't exist in theme to avoid unhelpful error.
+    * Added default templates for `external.html.twig`, `default.html.twig`, and `modular.html.twig`
 1. [](#bugfix)
     * Fixed `Security::sanitizeSVG()` creating an empty file if SVG file cannot be parsed
     * Fixed infinite loop in blueprints with `extend@` to a parent stream