+++ CHANGELOG.md
@@ -4,9 +4,11 @@
 1. [](#improved)
     * Fall back through various templates scenarios if they don't exist in theme to avoid unhelpful error.
     * Added default templates for `external.html.twig`, `default.html.twig`, and `modular.html.twig`
+    * Improve Media classes
 1. [](#bugfix)
     * Fixed `Security::sanitizeSVG()` creating an empty file if SVG file cannot be parsed
     * Fixed infinite loop in blueprints with `extend@` to a parent stream
+    * Added missing `Stream::create()` method
 
 # v1.7.0-rc.16
 ## 09/01/2020