+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ## mm/dd/2020
 
 1. [](#improved)
+    * Fall back through various templates scenarios if they don't exist in theme to avoid unhelpful error.
+    * Added default templates for `external.html.twig`, `default.html.twig`, and `modular.html.twig`
     * Improve Media classes
 1. [](#bugfix)
     * Fixed `Security::sanitizeSVG()` creating an empty file if SVG file cannot be parsed