+++ CHANGELOG.md
@@ -12,6 +12,7 @@
     * Added missing `onBlueprintCreated` event for Flex Pages
     * Fixed `onBlueprintCreated` firing multiple times recursively
     * Fixed media upload failing with custom folders
+    * Fixed `unset()` in `ObjectProperty` class
 
 # v1.7.0-rc.16
 ## 09/01/2020