+++ CHANGELOG.md
@@ -16,6 +16,7 @@
     * Fixed `FlexObject::freeMedia()` method causing media to become null
     * Fixed bug in `Flex Form` making it impossible to set nested values
     * Fixed `Flex User` avatar when using folder storage, also allow multiple images
+    * Fixed Referer reference during GPM calls.
 
 # v1.7.0-rc.16
 ## 09/01/2020